+++ examples/data-sources/fmc_continents/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_continents" "example" {
+  items = {
+    "Europe" = {
+    }
+  }
+}