+++ examples/data-sources/fmc_countries/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_countries" "example" {
+  items = {
+    "Poland" = {
+    }
+  }
+}