+++ examples/resources/fmc_geolocation/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_geolocation.example "<id>"