+++ examples/resources/fmc_service_access/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_service_access.example "<id>"