+++ internal/provider/model_fmc_geolocation.go
@@ -1,0 +1,297 @@
+// Copyright © 2023 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"context"
+	"fmt"
+	"slices"
+	"strconv"
+
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/tidwall/gjson"
+	"github.com/tidwall/sjson"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin types
+
+type Geolocation struct {
+	Id         types.String            `tfsdk:"id"`
+	Domain     types.String            `tfsdk:"domain"`
+	Name       types.String            `tfsdk:"name"`
+	Type       types.String            `tfsdk:"type"`
+	Continents []GeolocationContinents `tfsdk:"continents"`
+	Countries  []GeolocationCountries  `tfsdk:"countries"`
+}
+
+type GeolocationContinents struct {
+	Id types.Int64 `tfsdk:"id"`
+}
+
+type GeolocationCountries struct {
+	Id types.Int64 `tfsdk:"id"`
+}
+
+// End of section. //template:end types
+
+// Section below is generated&owned by "gen/generator.go". //template:begin minimumVersions
+
+// End of section. //template:end minimumVersions
+
+// Section below is generated&owned by "gen/generator.go". //template:begin getPath
+
+func (data Geolocation) getPath() string {
+	return "/api/fmc_config/v1/domain/{DOMAIN_UUID}/object/geolocations"
+}
+
+// End of section. //template:end getPath
+
+// Section below is generated&owned by "gen/generator.go". //template:begin toBody
+
+func (data Geolocation) toBody(ctx context.Context, state Geolocation) string {
+	body := ""
+	if data.Id.ValueString() != "" {
+		body, _ = sjson.Set(body, "id", data.Id.ValueString())
+	}
+	if !data.Name.IsNull() {
+		body, _ = sjson.Set(body, "name", data.Name.ValueString())
+	}
+	if len(data.Continents) > 0 {
+		body, _ = sjson.Set(body, "continents", []interface{}{})
+		for _, item := range data.Continents {
+			itemBody := ""
+			if !item.Id.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "id", item.Id.ValueInt64())
+			}
+			body, _ = sjson.SetRaw(body, "continents.-1", itemBody)
+		}
+	}
+	if len(data.Countries) > 0 {
+		body, _ = sjson.Set(body, "countries", []interface{}{})
+		for _, item := range data.Countries {
+			itemBody := ""
+			if !item.Id.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "id", item.Id.ValueInt64())
+			}
+			body, _ = sjson.SetRaw(body, "countries.-1", itemBody)
+		}
+	}
+	return body
+}
+
+// End of section. //template:end toBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBody
+
+func (data *Geolocation) fromBody(ctx context.Context, res gjson.Result) {
+	if value := res.Get("name"); value.Exists() {
+		data.Name = types.StringValue(value.String())
+	} else {
+		data.Name = types.StringNull()
+	}
+	if value := res.Get("type"); value.Exists() {
+		data.Type = types.StringValue(value.String())
+	} else {
+		data.Type = types.StringNull()
+	}
+	if value := res.Get("continents"); value.Exists() {
+		data.Continents = make([]GeolocationContinents, 0)
+		value.ForEach(func(k, res gjson.Result) bool {
+			parent := &data
+			data := GeolocationContinents{}
+			if value := res.Get("id"); value.Exists() {
+				data.Id = types.Int64Value(value.Int())
+			} else {
+				data.Id = types.Int64Null()
+			}
+			(*parent).Continents = append((*parent).Continents, data)
+			return true
+		})
+	}
+	if value := res.Get("countries"); value.Exists() {
+		data.Countries = make([]GeolocationCountries, 0)
+		value.ForEach(func(k, res gjson.Result) bool {
+			parent := &data
+			data := GeolocationCountries{}
+			if value := res.Get("id"); value.Exists() {
+				data.Id = types.Int64Value(value.Int())
+			} else {
+				data.Id = types.Int64Null()
+			}
+			(*parent).Countries = append((*parent).Countries, data)
+			return true
+		})
+	}
+}
+
+// End of section. //template:end fromBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBodyPartial
+
+// fromBodyPartial reads values from a gjson.Result into a tfstate model. It ignores null attributes in order to
+// uncouple the provider from the exact values that the backend API might summon to replace nulls. (Such behavior might
+// easily change across versions of the backend API.) For List/Set/Map attributes, the func only updates the
+// "managed" elements, instead of all elements.
+func (data *Geolocation) fromBodyPartial(ctx context.Context, res gjson.Result) {
+	if value := res.Get("name"); value.Exists() && !data.Name.IsNull() {
+		data.Name = types.StringValue(value.String())
+	} else {
+		data.Name = types.StringNull()
+	}
+	if value := res.Get("type"); value.Exists() && !data.Type.IsNull() {
+		data.Type = types.StringValue(value.String())
+	} else {
+		data.Type = types.StringNull()
+	}
+	for i := 0; i < len(data.Continents); i++ {
+		keys := [...]string{"id"}
+		keyValues := [...]string{strconv.FormatInt(data.Continents[i].Id.ValueInt64(), 10)}
+
+		parent := &data
+		data := (*parent).Continents[i]
+		parentRes := &res
+		var res gjson.Result
+
+		parentRes.Get("continents").ForEach(
+			func(_, v gjson.Result) bool {
+				found := false
+				for ik := range keys {
+					if v.Get(keys[ik]).String() != keyValues[ik] {
+						found = false
+						break
+					}
+					found = true
+				}
+				if found {
+					res = v
+					return false
+				}
+				return true
+			},
+		)
+		if !res.Exists() {
+			tflog.Debug(ctx, fmt.Sprintf("removing Continents[%d] = %+v",
+				i,
+				(*parent).Continents[i],
+			))
+			(*parent).Continents = slices.Delete((*parent).Continents, i, i+1)
+			i--
+
+			continue
+		}
+		if value := res.Get("id"); value.Exists() && !data.Id.IsNull() {
+			data.Id = types.Int64Value(value.Int())
+		} else {
+			data.Id = types.Int64Null()
+		}
+		(*parent).Continents[i] = data
+	}
+	for i := 0; i < len(data.Countries); i++ {
+		keys := [...]string{"id"}
+		keyValues := [...]string{strconv.FormatInt(data.Countries[i].Id.ValueInt64(), 10)}
+
+		parent := &data
+		data := (*parent).Countries[i]
+		parentRes := &res
+		var res gjson.Result
+
+		parentRes.Get("countries").ForEach(
+			func(_, v gjson.Result) bool {
+				found := false
+				for ik := range keys {
+					if v.Get(keys[ik]).String() != keyValues[ik] {
+						found = false
+						break
+					}
+					found = true
+				}
+				if found {
+					res = v
+					return false
+				}
+				return true
+			},
+		)
+		if !res.Exists() {
+			tflog.Debug(ctx, fmt.Sprintf("removing Countries[%d] = %+v",
+				i,
+				(*parent).Countries[i],
+			))
+			(*parent).Countries = slices.Delete((*parent).Countries, i, i+1)
+			i--
+
+			continue
+		}
+		if value := res.Get("id"); value.Exists() && !data.Id.IsNull() {
+			data.Id = types.Int64Value(value.Int())
+		} else {
+			data.Id = types.Int64Null()
+		}
+		(*parent).Countries[i] = data
+	}
+}
+
+// End of section. //template:end fromBodyPartial
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBodyUnknowns
+
+// fromBodyUnknowns updates the Unknown Computed tfstate values from a JSON.
+// Known values are not changed (usual for Computed attributes with UseStateForUnknown or with Default).
+func (data *Geolocation) fromBodyUnknowns(ctx context.Context, res gjson.Result) {
+	if data.Type.IsUnknown() {
+		if value := res.Get("type"); value.Exists() {
+			data.Type = types.StringValue(value.String())
+		} else {
+			data.Type = types.StringNull()
+		}
+	}
+}
+
+// End of section. //template:end fromBodyUnknowns
+
+// Section below is generated&owned by "gen/generator.go". //template:begin Clone
+
+// End of section. //template:end Clone
+
+// Section below is generated&owned by "gen/generator.go". //template:begin toBodyNonBulk
+
+// End of section. //template:end toBodyNonBulk
+
+// Section below is generated&owned by "gen/generator.go". //template:begin findObjectsToBeReplaced
+
+// End of section. //template:end findObjectsToBeReplaced
+
+// Section below is generated&owned by "gen/generator.go". //template:begin clearItemIds
+
+// End of section. //template:end clearItemIds
+
+// Section below is generated&owned by "gen/generator.go". //template:begin toBodyPutDelete
+
+// End of section. //template:end toBodyPutDelete
+
+// Section below is generated&owned by "gen/generator.go". //template:begin adjustBody
+
+// End of section. //template:end adjustBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin adjustBodyBulk
+
+// End of section. //template:end adjustBodyBulk