+++ internal/provider/model_fmc_service_access.go
@@ -1,0 +1,288 @@
+// Copyright © 2023 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"context"
+	"fmt"
+	"slices"
+
+	"github.com/hashicorp/go-version"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/tidwall/gjson"
+	"github.com/tidwall/sjson"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin types
+
+type ServiceAccess struct {
+	Id            types.String         `tfsdk:"id"`
+	Domain        types.String         `tfsdk:"domain"`
+	Name          types.String         `tfsdk:"name"`
+	Type          types.String         `tfsdk:"type"`
+	DefaultAction types.String         `tfsdk:"default_action"`
+	Rules         []ServiceAccessRules `tfsdk:"rules"`
+}
+
+type ServiceAccessRules struct {
+	Action             types.String                           `tfsdk:"action"`
+	GeolocationSources []ServiceAccessRulesGeolocationSources `tfsdk:"geolocation_sources"`
+}
+
+type ServiceAccessRulesGeolocationSources struct {
+	Id   types.String `tfsdk:"id"`
+	Type types.String `tfsdk:"type"`
+}
+
+// End of section. //template:end types
+
+// Section below is generated&owned by "gen/generator.go". //template:begin minimumVersions
+var minFMCVersionServiceAccess = version.Must(version.NewVersion("7.7"))
+
+// End of section. //template:end minimumVersions
+
+// Section below is generated&owned by "gen/generator.go". //template:begin getPath
+
+func (data ServiceAccess) getPath() string {
+	return "/api/fmc_config/v1/domain/{DOMAIN_UUID}/object/serviceaccessobjects"
+}
+
+// End of section. //template:end getPath
+
+// Section below is generated&owned by "gen/generator.go". //template:begin toBody
+
+func (data ServiceAccess) toBody(ctx context.Context, state ServiceAccess) string {
+	body := ""
+	if data.Id.ValueString() != "" {
+		body, _ = sjson.Set(body, "id", data.Id.ValueString())
+	}
+	if !data.Name.IsNull() {
+		body, _ = sjson.Set(body, "name", data.Name.ValueString())
+	}
+	body, _ = sjson.Set(body, "type", "ServiceAccessObject")
+	if !data.DefaultAction.IsNull() {
+		body, _ = sjson.Set(body, "defaultAction", data.DefaultAction.ValueString())
+	}
+	if len(data.Rules) > 0 {
+		body, _ = sjson.Set(body, "rules", []interface{}{})
+		for _, item := range data.Rules {
+			itemBody := ""
+			if !item.Action.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "action", item.Action.ValueString())
+			}
+			if len(item.GeolocationSources) > 0 {
+				itemBody, _ = sjson.Set(itemBody, "geoSources", []interface{}{})
+				for _, childItem := range item.GeolocationSources {
+					itemChildBody := ""
+					if !childItem.Id.IsNull() {
+						itemChildBody, _ = sjson.Set(itemChildBody, "id", childItem.Id.ValueString())
+					}
+					if !childItem.Type.IsNull() {
+						itemChildBody, _ = sjson.Set(itemChildBody, "type", childItem.Type.ValueString())
+					}
+					itemBody, _ = sjson.SetRaw(itemBody, "geoSources.-1", itemChildBody)
+				}
+			}
+			body, _ = sjson.SetRaw(body, "rules.-1", itemBody)
+		}
+	}
+	return body
+}
+
+// End of section. //template:end toBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBody
+
+func (data *ServiceAccess) fromBody(ctx context.Context, res gjson.Result) {
+	if value := res.Get("name"); value.Exists() {
+		data.Name = types.StringValue(value.String())
+	} else {
+		data.Name = types.StringNull()
+	}
+	if value := res.Get("type"); value.Exists() {
+		data.Type = types.StringValue(value.String())
+	} else {
+		data.Type = types.StringNull()
+	}
+	if value := res.Get("defaultAction"); value.Exists() {
+		data.DefaultAction = types.StringValue(value.String())
+	} else {
+		data.DefaultAction = types.StringNull()
+	}
+	if value := res.Get("rules"); value.Exists() {
+		data.Rules = make([]ServiceAccessRules, 0)
+		value.ForEach(func(k, res gjson.Result) bool {
+			parent := &data
+			data := ServiceAccessRules{}
+			if value := res.Get("action"); value.Exists() {
+				data.Action = types.StringValue(value.String())
+			} else {
+				data.Action = types.StringNull()
+			}
+			if value := res.Get("geoSources"); value.Exists() {
+				data.GeolocationSources = make([]ServiceAccessRulesGeolocationSources, 0)
+				value.ForEach(func(k, res gjson.Result) bool {
+					parent := &data
+					data := ServiceAccessRulesGeolocationSources{}
+					if value := res.Get("id"); value.Exists() {
+						data.Id = types.StringValue(value.String())
+					} else {
+						data.Id = types.StringNull()
+					}
+					if value := res.Get("type"); value.Exists() {
+						data.Type = types.StringValue(value.String())
+					} else {
+						data.Type = types.StringNull()
+					}
+					(*parent).GeolocationSources = append((*parent).GeolocationSources, data)
+					return true
+				})
+			}
+			(*parent).Rules = append((*parent).Rules, data)
+			return true
+		})
+	}
+}
+
+// End of section. //template:end fromBody
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBodyPartial
+
+// fromBodyPartial reads values from a gjson.Result into a tfstate model. It ignores null attributes in order to
+// uncouple the provider from the exact values that the backend API might summon to replace nulls. (Such behavior might
+// easily change across versions of the backend API.) For List/Set/Map attributes, the func only updates the
+// "managed" elements, instead of all elements.
+func (data *ServiceAccess) fromBodyPartial(ctx context.Context, res gjson.Result) {
+	if value := res.Get("name"); value.Exists() && !data.Name.IsNull() {
+		data.Name = types.StringValue(value.String())
+	} else {
+		data.Name = types.StringNull()
+	}
+	if value := res.Get("type"); value.Exists() && !data.Type.IsNull() {
+		data.Type = types.StringValue(value.String())
+	} else {
+		data.Type = types.StringNull()
+	}
+	if value := res.Get("defaultAction"); value.Exists() && !data.DefaultAction.IsNull() {
+		data.DefaultAction = types.StringValue(value.String())
+	} else {
+		data.DefaultAction = types.StringNull()
+	}
+	{
+		l := len(res.Get("rules").Array())
+		tflog.Debug(ctx, fmt.Sprintf("rules array resizing from %d to %d", len(data.Rules), l))
+		for i := len(data.Rules); i < l; i++ {
+			data.Rules = append(data.Rules, ServiceAccessRules{})
+		}
+		if len(data.Rules) > l {
+			data.Rules = data.Rules[:l]
+		}
+	}
+	for i := range data.Rules {
+		parent := &data
+		data := (*parent).Rules[i]
+		parentRes := &res
+		res := parentRes.Get(fmt.Sprintf("rules.%d", i))
+		if value := res.Get("action"); value.Exists() && !data.Action.IsNull() {
+			data.Action = types.StringValue(value.String())
+		} else {
+			data.Action = types.StringNull()
+		}
+		for i := 0; i < len(data.GeolocationSources); i++ {
+			keys := [...]string{"id", "type"}
+			keyValues := [...]string{data.GeolocationSources[i].Id.ValueString(), data.GeolocationSources[i].Type.ValueString()}
+
+			parent := &data
+			data := (*parent).GeolocationSources[i]
+			parentRes := &res
+			var res gjson.Result
+
+			parentRes.Get("geoSources").ForEach(
+				func(_, v gjson.Result) bool {
+					found := false
+					for ik := range keys {
+						if v.Get(keys[ik]).String() != keyValues[ik] {
+							found = false
+							break
+						}
+						found = true
+					}
+					if found {
+						res = v
+						return false
+					}
+					return true
+				},
+			)
+			if !res.Exists() {
+				tflog.Debug(ctx, fmt.Sprintf("removing GeolocationSources[%d] = %+v",
+					i,
+					(*parent).GeolocationSources[i],
+				))
+				(*parent).GeolocationSources = slices.Delete((*parent).GeolocationSources, i, i+1)
+				i--
+
+				continue
+			}
+			if value := res.Get("id"); value.Exists() && !data.Id.IsNull() {
+				data.Id = types.StringValue(value.String())
+			} else {
+				data.Id = types.StringNull()
+			}
+			if value := res.Get("type"); value.Exists() && !data.Type.IsNull() {
+				data.Type = types.StringValue(value.String())
+			} else {
+				data.Type = types.StringNull()
+			}
+			(*parent).GeolocationSources[i] = data
+		}
+		(*parent).Rules[i] = data
+	}
+}
+
+// End of section. //template:end fromBodyPartial
+
+// Section below is generated&owned by "gen/generator.go". //template:begin fromBodyUnknowns
+
+// fromBodyUnknowns updates the Unknown Computed tfstate values from a JSON.
+// Known values are not changed (usual for Computed attributes with UseStateForUnknown or with Default).
+func (data *ServiceAccess) fromBodyUnknowns(ctx context.Context, res gjson.Result) {
+	if data.Type.IsUnknown() {
+		if value := res.Get("type"); value.Exists() {
+			data.Type = types.StringValue(value.String())
+		} else {
+			data.Type = types.StringNull()
+		}
+	}
+}
+
+// End of section. //template:end fromBodyUnknowns
+
+func (data ServiceAccess) adjustBody(ctx context.Context, req string) string {
+
+	// Add sequence numbers to the entities
+	for i := range len(data.Rules) {
+		req, _ = sjson.Set(req, fmt.Sprintf("rules.%d.sequenceNumber", i), i+1)
+	}
+
+	return req
+}