+++ docs/guides/changelog.md
@@ -7,6 +7,12 @@ description: |-
 
 # Changelog
 
+## 2.0.0-rc5 (Unreleased)
+
+- (Enhancement) Add support for `fmc_countries` and `fmc_continents` data sources
+- (Enhancement) Add support for `fmc_geolocation` resource and data source
+- (Enhancement) Add support for `fmc_service_access` resource and data source
+
 ## 2.0.0-rc4
 
 - (Fix) Corrected URL encoding for multiple resources